configotus/requirements.txt
google-cloud-bigquery
firebase-admin
google-cloud-firestore
google-cloud-iot
google-cloud-pubsub
protobuf
